feat: move body updation to new hook
